rename: grokfeed → devskim
Rename Python package, CLI entrypoint, config dir (~/.devskim), app
class, all imports, test patch paths, CI/release workflows, demo
script, and README. Avoids xAI Grok trademark overlap.

PyPI tombstone release and GitHub/Homebrew repo renames are manual
follow-up steps documented in the PR description.
