+++ .github/workflows/ci.yml
@@ -16,7 +16,7 @@ jobs:
       - run: uv sync
       - run: uv run ruff check .
       - run: uv run ruff format --check .
-      - run: uv run mypy grokfeed
+      - run: uv run mypy devskim
 
   test:
     runs-on: ${{ matrix.os }}
@@ -31,7 +31,7 @@ jobs:
         with:
           python-version: ${{ matrix.python }}
       - run: uv sync
-      - run: uv run pytest --cov=grokfeed --cov-report=xml --cov-report=term-missing
+      - run: uv run pytest --cov=devskim --cov-report=xml --cov-report=term-missing
       - name: Upload coverage
         if: matrix.os == 'ubuntu-latest' && matrix.python == '3.12'
         uses: codecov/codecov-action@v4
