+++ .github/workflows/demo.yml
@@ -27,7 +27,7 @@ jobs:
           curl -fsSL "https://github.com/charmbracelet/vhs/releases/download/v${VHS_VERSION}/vhs_${VHS_VERSION}_amd64.deb" \
             -o vhs.deb && sudo dpkg -i vhs.deb
 
-      - name: Install grokfeed
+      - name: Install devskim
         run: uv tool install .
 
       - name: Record demo
